fix for round

+++ flappy_bird.py
@@ -195,8 +195,8 @@ class Pipe():
         bird_mask = bird.get_mask()
         top_mask = pygame.mask.from_surface(self.PIPE_TOP)
         bottom_mask = pygame.mask.from_surface(self.PIPE_BOTTOM)
-        top_offset = (self.x - bird.x, self.top - round(bird.y))
-        bottom_offset = (self.x - bird.x, self.bottom - round(bird.y))
+        top_offset = (self.x - bird.x, self.top - int(round(bird.y)))
+        bottom_offset = (self.x - bird.x, self.bottom - int(round(bird.y)))
 
         b_point = bird_mask.overlap(bottom_mask, bottom_offset)
         t_point = bird_mask.overlap(top_mask,top_offset)
